requirements.txt
django==4.0.2
djangorestframework==3.13.1
pytest==7.0.1